+++ src/settings/MappingProfiles/detailsSections/edit/InvoiceDetailSection/ExtendedInformation.js
@@ -5,7 +5,6 @@ import {
   FormattedMessage,
 } from 'react-intl';
 import { Field } from 'redux-form';
-import { noop } from 'lodash';
 
 import {
   Accordion,
@@ -27,6 +26,7 @@ import { TRANSLATION_ID_PREFIX } from '../../constants';
 import {
   okapiShape,
   createOptionsList,
+  validateRequiredField,
   BOOLEAN_ACTIONS,
 } from '../../../../../utils';
 
@@ -69,7 +69,8 @@ export const ExtendedInformation = ({
             isRemoveValueAllowed
             wrapperLabel={`${TRANSLATION_ID_PREFIX}.wrapper.acceptedValues`}
             acceptedValuesList={paymentMethodsList}
-            validation={noop}
+            validation={validateRequiredField}
+            required
             okapi={okapi}
           />
         </Col>
@@ -105,7 +106,8 @@ export const ExtendedInformation = ({
             isRemoveValueAllowed
             wrapperLabel={`${TRANSLATION_ID_PREFIX}.wrapper.acceptedValues`}
             acceptedValuesList={currenciesList}
-            validation={noop}
+            validation={validateRequiredField}
+            required
             okapi={okapi}
           />
         </Col>